refact: moved konva files out of test dir
