+++ frontend/components/DocumentCanvas/DrawingArea.tsx
